isolate inference engine from the UI in frontend and backend
Frontend:
- Move Ollama/OpenAI provider clients out of the UI layer.
- Introduce a neutral facade (src/lib/apis/inference.ts) that re-exports the
  engine implementations (src/lib/inference/{ollama,openai}.ts).
- Repoint all 13 UI consumers at the facade; no markup/style/logic changes.

Backend:
- Move the inference engine out of routers: open_webui/inference/{ollama,openai}.py
  now own the implementation, exposed through the inference facade package.
- routers/{ollama,openai}.py become thin HTTP shims that re-export the engine
  (keeps /ollama and /openai endpoints, and the UI, working unchanged).
- main.py, utils/chat, utils/embeddings, utils/models, routers/pipelines now
  depend on open_webui.inference instead of routers.
