+++ backend/open_webui/routers/pipelines.py
@@ -21,7 +21,7 @@ from open_webui.constants import ERROR_MESSAGES
 from open_webui.env import AIOHTTP_CLIENT_SESSION_SSL, AIOHTTP_FILE_STREAM_CHUNK_SIZE
 from open_webui.events import EVENTS, publish_event
 from open_webui.models.config import Config
-from open_webui.routers.openai import get_all_models_responses
+from open_webui.inference.openai import get_all_models_responses
 from open_webui.utils.auth import get_admin_user
 from pydantic import BaseModel
 from starlette.responses import FileResponse
