+++ backend/open_webui/utils/chat.py
@@ -13,10 +13,10 @@ from open_webui.env import BYPASS_MODEL_ACCESS_CONTROL, GLOBAL_LOG_LEVEL
 from open_webui.functions import generate_function_chat_completion
 from open_webui.models.models import Models
 from open_webui.models.users import UserModel
-from open_webui.routers.ollama import (
+from open_webui.inference.ollama import (
     generate_chat_completion as generate_ollama_chat_completion,
 )
-from open_webui.routers.openai import (
+from open_webui.inference.openai import (
     generate_chat_completion as generate_openai_chat_completion,
 )
 from open_webui.routers.pipelines import (
