+++ src/lib/components/AddConnectionModal.svelte
@@ -3,8 +3,7 @@
 	import { getContext, onMount } from 'svelte';
 	const i18n = getContext('i18n');
 
-	import { verifyOpenAIConnection } from '$lib/apis/openai';
-	import { verifyOllamaConnection } from '$lib/apis/ollama';
+	import { verifyOpenAIConnection, verifyOllamaConnection } from '$lib/apis/inference';
 
 	import Modal from '$lib/components/common/Modal.svelte';
 	import Plus from '$lib/components/icons/Plus.svelte';
