+++ src/lib/components/admin/Settings/Models/ManageModelsModal.svelte
@@ -10,7 +10,7 @@
 	import XMark from '$lib/components/icons/XMark.svelte';
 	import Modal from '$lib/components/common/Modal.svelte';
 	import ManageOllama from './Manage/ManageOllama.svelte';
-	import { getOllamaConfig } from '$lib/apis/ollama';
+	import { getOllamaConfig } from '$lib/apis/inference';
 	import Spinner from '$lib/components/common/Spinner.svelte';
 	import ManageMultipleOllama from './Manage/ManageMultipleOllama.svelte';
