+++ src/lib/components/playground/Chat.svelte
@@ -15,7 +15,7 @@
 	} from '$lib/constants';
 	import { WEBUI_NAME, config, user, models, settings } from '$lib/stores';
 
-	import { chatCompletion } from '$lib/apis/openai';
+	import { chatCompletion } from '$lib/apis/inference';
 
 	import { splitStream } from '$lib/utils';
 	import Collapsible from '../common/Collapsible.svelte';
